New cogs update!
The update clearly has an awesome feature, with more file organization. I am hinting that in the next update I will have mysql database for economy and make a hosting.md

+++ TheImperialGod/cogs/math.py
@@ -13,8 +13,8 @@ class Math(commands.Cog):
     
     @commands.command()
     async def add(self, ctx, num1, num2):
-        num1 = float(num1)
-        num2 = float(num2)
+        num1 = int(num1)
+        num2 = int(num2)
         res = num1 + num2
         embed = discord.Embed(title = "Sum:", color = ctx.author.color)
         embed.add_field(name = "Result:", value = f"`{res}`")
@@ -22,8 +22,8 @@ class Math(commands.Cog):
 
     @commands.command()
     async def subtract(self, ctx, num1, num2):
-        num1 = float(num1)
-        num2 = float(num2)
+        num1 = int(num1)
+        num2 = int(num2)
         res = num1 - num2
         embed = discord.Embed(title = "Difference:", color = ctx.author.color)
         embed.add_field(name = "Result:", value = f"`{res}`")
@@ -31,8 +31,8 @@ class Math(commands.Cog):
 
     @commands.command()
     async def multiply(self, ctx, num1, num2):
-        num1 = float(num1)
-        num2 = float(num2)
+        num1 = int(num1)
+        num2 = int(num2)
         res = num1 * num2
         embed = discord.Embed(title = "Product:", color = ctx.author.color)
         embed.add_field(name = "Result:", value = f"`{res}`")
@@ -40,8 +40,8 @@ class Math(commands.Cog):
 
     @commands.command()
     async def divide(self, ctx, num1, num2):
-        num1 = float(num1)
-        num2 = float(num2)
+        num1 = int(num1)
+        num2 = int(num2)
         res = num1 / num2
         embed = discord.Embed(title = "Quotient:", color = ctx.author.color)
         embed.add_field(name = "Result:", value = f"`{res}`")
@@ -49,7 +49,7 @@ class Math(commands.Cog):
     
     @commands.command()
     async def square(self, ctx, num1):
-        num1 = float(num1)
+        num1 = int(num1)
         res = num1 * num1
         embed = discord.Embed(title = "Square:", color = ctx.author.color)
         embed.add_field(name = "Result:", value = f"`{res}`")
@@ -57,7 +57,7 @@ class Math(commands.Cog):
     
     @commands.command()
     async def sqrt(self, ctx, num1):
-        num1 = float(num1)
+        num1 = int(num1)
         res = math.sqrt(num1)
         
         embed = discord.Embed(title = "Sum:", color = ctx.author.color)
@@ -66,8 +66,8 @@ class Math(commands.Cog):
     
     @commands.command()
     async def pow(self, ctx, num1, num2):
-        num1 = float(num1)
-        num2 = float(num2)
+        num1 = int(num1)
+        num2 = int(num2)
         res = math.pow(num1, num2)
 
         embed = discord.Embed(title = "Sum:", color = ctx.author.color)
